+++ ConfigurationExtractor/PropertyConfigurationExtractor.php
@@ -16,6 +16,8 @@ use Doctrine\Common\Annotations\AnnotationException;
 use Doctrine\Common\Annotations\AnnotationReader;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\ParamConverter;
 use Symfony\Component\Validator\Constraints\All;
+use Symfony\Component\Validator\Constraints\Date;
+use Symfony\Component\Validator\Constraints\DateTime;
 use Symfony\Component\Validator\Constraints\NotBlank;
 use Symfony\Component\Validator\Constraints\NotNull;
 use Symfony\Component\Validator\Constraints\Type;
@@ -76,6 +78,10 @@ class PropertyConfigurationExtractor
         $arrayAnnotation = $annotationReader->getPropertyAnnotation($property, All::class);
         /** @var Type $typeAnnotation */
         $typeAnnotation = $annotationReader->getPropertyAnnotation($property, Type::class);
+        /** @var DateTime $dateTimeAnnotation */
+        $dateTimeAnnotation = $annotationReader->getPropertyAnnotation($property, DateTime::class);
+        /** @var Date $dateAnnotation */
+        $dateAnnotation = $annotationReader->getPropertyAnnotation($property, Date::class);
         /** @var MapTo $mapToAnnotation */
         $mapToAnnotation = $annotationReader->getPropertyAnnotation($property, MapTo::class);
         /** @var NotNull $notNullAnnotation */
@@ -91,6 +97,10 @@ class PropertyConfigurationExtractor
 
         if ($this->isCollection) {
              $typeAnnotation = $this->findTypeConstraint($arrayAnnotation) ?? $typeAnnotation;
+        }
+
+        if ($dateTimeAnnotation !== null || $dateAnnotation !== null) {
+            $this->type = \DateTime::class;
         }
 
         if ($typeAnnotation !== null && \class_exists($typeAnnotation->type)) {